Fix bug in 'calc_tile_map'
The index we access in maps 3d-array was wrong. We indexes it by: tile_row_rum * TILE_COUNT + tile_col_num + tid + i * numThreads
however, we did not take into consideration that the array is 3d, and we are missing the dimension of the Colors.
fixed it by multiplying the first 2 elements by COLOR_VALUES